Macro to iterate over nodes of known type
There is a pattern in the codebase when a list of nodes is iterated and
the type of each of the nodes is equal and known. This is for example to
iterate over lists of pins, etc.

This introduces a macro ASTNODE_ITERATE that has three parameters:

 - type: Is the type of the nodes (Ast<type> is the matching class name)

 - it: Name of the iterator variable

 - init: Initialization statement for iterator

All occurrences of this pattern have been accordingly replaced:

for (Ast<type> *<it> = <init>; <it>; <it>=<it>->nextp()->cast<type>())

This improves readability and some time in the future this may be a good
marker to replace data containers or similar.