Wait for final parameter values before adding them
Most local parameters have their values set in AllModules tree, but for
some the value can be overriden in TopModules. This happens i.e. when
they depend on a different parameter that is overridden in Surelog
command line.

The parameters are tracked per module (or package, scope etc.) and only
their final values are added to the tree. Non-local parameters are still
added as pins to their cells.

Signed-off-by: Rafal Kapuscik <rkapuscik@antmicro.com>